dma-buf: correct printf format for counter.
It needs a typecast, since the f_count.counter is defined as int on ARM
and long on x86.

+++ drivers/base/dma-buf.c
@@ -598,10 +598,10 @@ static int dma_buf_describe(struct seq_file *s)
 
 		seq_printf(s, "\t");
 
-		seq_printf(s, "\t%s\t%08zu\t%08x\t%08x\t%08d\n",
+		seq_printf(s, "\t%s\t%08zu\t%08x\t%08x\t%08ld\n",
 				buf_obj->exp_name, buf_obj->size,
 				buf_obj->file->f_flags, buf_obj->file->f_mode,
-				buf_obj->file->f_count.counter);
+				(long)(buf_obj->file->f_count.counter));
 
 		seq_printf(s, "\t\tAttached Devices:\n");
 		attach_count = 0;
